Revert back a ways preparing to use a different strategy to eliminate ifdefs.
