+++ server/hw/hw_gfx.c
@@ -234,7 +234,6 @@ INLINE void X11_Mogrify(dat x, dat y, uldat len) {
 }
 
 #include "hw_x/util.h"
-#include "hw_x/x11_utf32_to_charset.c"
 
 #undef XDRAW_ANY
