--- server/hw/hw_xft/x11_utf32_to_charset.c
@@ -1,5 +1,0 @@
-/* xft version of X11_UTF_32_to_charset_function */
-static Tutf_function X11_UTF_32_to_charset_function(CONST byte *charset) {
-    // this is sufficient for xft fonts which are 16-bit unicode
-    return X11_UTF_32_to_UCS_2;
-}
