+++ server/hw/hw_xft/x11setcolors.h
